-version bump to 0.10
-Changed mutation type JCheckBox code - boxes now generated based on the
mutation types observed.  Filtering is linked to boxes by the same data being
sorted the same way to set up the boxes (and the order of the filter mask) and
the order the filter mask is read.

-Allow for compound het filtering if CD_Pred column missing (really, test for presence
of all column headers)

-Removing old lines that were commented out due to non-use or improvement
-Continued code clean-up

-Removed deprecated VarData methods countAffNorm() and isMendFilt()

-fixed recently introduced File open bugs (had to reinit Frame, either by
calling again, or by calling frameInit() in main class.

-fixed memory leak, in which opening a new file didn't really release memory
of old file!!  Not sure when it was introduced, but it seems to be fixed now.

-Some minor label changes


##OLD

-0.9b_BETA
-can now read in a list of mutation types, in column 'type'.  Delimiter String
defined in MultiStringMapper constructor.
Can appropriately filter mutation types, and save them (although order is
altered in save.)

-code simplification - moved test for Mapper entry presence to addData(Object)
methods - removed multiple instances from VarData

-added error message and recognition when jdk not present.

-removed Hiding of value '301', which was originally implemented to hide
 CD_Pred scores that were NA (but which cannot be displayed as NA for the
 column to stay as int.) CD_Pred 'NA' values will remain as 301, but won't be
 hidden.

-All Strings center aligned, numbers right aligned.

-fixed some bugs with Compound Het Filtering - exceptions when trying to
filter on variant or gene that is not mendHetRec. Now shows an error dialog
instead of throwing an exception.